+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.15.0
+## 0.15.0 (September 28, 2016)
 
 ### Enhancements
 - **Annual Changes**